modifications for NIHVIVO-246.
changed the getSuperPropertyURIs and getSuperClassURIs methods in the property and class DaoJenas (and the interfaces that they implement, and the filtering code) to take an additional boolean argument named direct and to pass it to the Jena call.

changed the edit controllers to invoke the above methods with direct=false.

changed the recursive methods within the daos to invoke the above methods with direct=true