VIVO-87 First attempt at a caching filter.
Appears to work with Firefox cache and Chrome cache, except for language-dependent stuff.

Is that because it isn't configured correctly, or is it because Firefox and Chrome aren't attempting to keep more than one version of a page in their caches? Even if they do see a "Vary: Accept-Language" header?