VIVO-118 Make getDataPropertyByURI() an unfiltered operation.
This method is used to determine whether a given property is a data property or not. If the user is restricted from displaying this property, then the filtered answer is NO. But that is really inappropriate. The real questions are whether the user can see the values of this property, and whether the user can add or modify values of this property -- not whether the user can see or modify the property itself.