Preserve original when applying ASCIIFoldingFilter
This is a suggestion, but here at Duke, we recently ran into a situation where a faculty member with an umlat (`ö`) in her name could be found only with her name spelled out with `o`. This change allows her name to be searched either way. This seems like the appropriate default behavior.