VIVO-766 This comparator is invalid under Java 8.
Because subclass/superclass is only a partial ordering, this breaks the contract for Comparator,
and throws an IllegalArgumentException under Java 8. In particular, for classes sub, super and other,
we have sub=other, sub<super, which should imply other<super, but instead other=super.

As far as I can determine, this sort is never used anyway, so there should be no impact in removing it.
PropertyInstanceDaoJena sorts these results again before using them, as does GroupedPropertyList.