VIVO-1031 Add seeding to the cache, and do some cleanup.
We use one CONSTRUCT query to find the labels and types of all individuals that are related to our subject,
and seed the cache with IndividualSDB instance formed from that information. This is faster than creating
such instances one a a time.

Faster, that is, for individuals with large profiles. When using SDB.