Fix RPC Endpoint test to actual test against RPC Endpoint and not REST Endpoint.
The static mocks must be recreated on each run or tests will fail (probably due to the design ramifications of SimpleMock).
Most of the mocks had to be converted into static variables.

To make sure that the action.run() is actually called by explicitly calling verify().

The staging branch is merged into this branch to accommodate breaking changes.