Add getMethod() and rpcMethodUnsupported(), also updating tests accordingly.
The method needs to be detected and appropriately return "501 Not Implemented".
Add getMethod() to OperationData to facilitate this.
Add rpcMethodUnsupported() to Action and utilize rpcMethodUnsupported() to validate the HTTP Method.

Some RPCEndpointTest tests are updated to accomodate this.
Remove some unecessary steps from RPCEndpointTest, as the `@Mock` already handles mock instantiation.