Properly handle unimplemented Methods in RPC and replace testing framework.
The RPC endpoints should return 501 for the GET, PUT, and DELETE methods.

The tests are updated to handle this.
Using EasyMock continued to be enough problems that I feel that it is necessary to replace EasyMock (and the recently added PowerMock) with Mockito.

Mockito allows for much cleaner test writing and provides a more stable API.

The other projects depending on EasyMock either use it in very few ways or are using it when they should instead be using JUnit (specifically the JUnit assertions).