Implement Argument Matcher and test the OpenAPI documentation for RPC.
The OpenAPIv3 requires the "info" with "title" and "version".
This provides a temporary hardcoded values for these and I plan on having a discussion on what to do in this regard.

Utilize the Mockito ArgumentMatcher API to write a custom test for ensuring that the generated JSON and YAML OpenAPI Specs match pre-defined expecations.

The REST is stubbed out but not implemented within the scope of this commit.