Only a single type of test data is needed.
It just occurred to me that this seems silly.
We are testing against the OpenAPI class and so the source data to compare against only needs to be in a single type.
I chose json over yaml.

Change the behavior of the program to just use a single test file for each expected response.