Issue 3751: Remove ICU4J dependency.
This is only used in a handful of places.
The core java libraries provide the same functionality but with a less exchaustive locale database.

The IBM ICU is more up to date than the Java version but is otherwise identical.

see: https://unicode-org.github.io/icu/userguide/icu4j/#introduction-to-icu4j