VIVO 3838: Add back the `@Override` annotations.
The `@Override` annotation is not needed anymore by Java.
I historically want to remove it for this reason as it saves space and makes the code a little easier to maintain.

The migration between Jena 3 to Jena 4 presented me with a strong reason to not only keep these annotations but to also promote their continued use.

When upgrading and the dependent model completely removes the overridden method, the `@Override` results in an error.
This makes it easy to identify which methods are removed during such a migration.
If the removed method is to be still maintained, then the `@Override` can simply be removed.